smart-food-ordering-platform/user-service/requirements.txt
flask==3.0.0
flask-cors==4.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pyjwt==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0
